MNT: rearrange to make iocmanager stuff reusable
